backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# GitHub Models / OpenAI SDK
openai>=1.30.0

# Resume Parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
